+++ client/examples/offline_signing/2_transaction_signing.rs
@@ -32,7 +32,7 @@ async fn main() -> Result<()> {
 
     // Signs the prepared transaction offline.
     let unlocks = secret_manager
-        .sign_transaction_essence(&prepared_transaction_data)
+        .sign_transaction_essence(&prepared_transaction_data, None)
         .await?;
     let signed_transaction = TransactionPayload::new(prepared_transaction_data.essence.clone(), unlocks)?;