+++ client/src/api/block_builder/input_selection/automatic.rs
@@ -18,12 +18,13 @@ use iota_types::{
 use crate::{
     api::{
         block_builder::input_selection::core::{Error as InputSelectionError, InputSelection, Selected},
+        input_selection::is_alias_transition,
         ClientBlockBuilder, ADDRESS_GAP_RANGE,
     },
     constants::HD_WALLET_TYPE,
     node_api::indexer::query_parameters::QueryParameter,
     secret::types::InputSigningData,
-    Error, Result,
+    unix_timestamp_now, Error, Result,
 };
 
 impl<'a> ClientBlockBuilder<'a> {
@@ -49,12 +50,7 @@ impl<'a> ClientBlockBuilder<'a> {
                     QueryParameter::HasExpiration(true),
                     QueryParameter::HasStorageDepositReturn(false),
                     // Ignore outputs that aren't expired yet
-                    QueryParameter::ExpiresBefore(
-                        instant::SystemTime::now()
-                            .duration_since(instant::SystemTime::UNIX_EPOCH)
-                            .expect("time went backwards")
-                            .as_secs() as u32,
-                    ),
+                    QueryParameter::ExpiresBefore(unix_timestamp_now()),
                 ])
                 .await?,
         );
@@ -77,6 +73,7 @@ impl<'a> ClientBlockBuilder<'a> {
 
         // First get inputs for utxo chains (Alias, Foundry, NFT outputs).
         let mut available_inputs = self.get_utxo_chains_inputs(self.outputs.iter()).await?;
+
         let required_inputs_for_sender_or_issuer = self.get_inputs_for_sender_and_issuer(&available_inputs).await?;
         let required_inputs_for_sender_or_issuer_ids = required_inputs_for_sender_or_issuer
             .iter()
@@ -87,13 +84,21 @@ impl<'a> ClientBlockBuilder<'a> {
         available_inputs.sort_unstable_by_key(|input| *input.output_id());
         available_inputs.dedup_by_key(|input| *input.output_id());
 
-        // Assume that we own the addresses for inputs that are required for the provided outputs
-        let mut available_input_addresses = available_inputs
-            .iter()
-            .map(|input| Ok(Address::try_from_bech32(&input.bech32_address)?.1))
-            .collect::<Result<Vec<Address>>>()?;
-
         let current_time = self.client.get_time_checked().await?;
+        // Assume that we own the addresses for inputs that are required for the provided outputs
+        let mut available_input_addresses = Vec::new();
+        for input in &available_inputs {
+            let alias_transition = is_alias_transition(input, &self.outputs);
+            let (required_unlock_address, unlocked_alias_or_nft_address) = input.output.required_and_unlocked_address(
+                current_time,
+                input.output_id(),
+                alias_transition.map(|(alias_transition, _)| alias_transition),
+            )?;
+            available_input_addresses.push(required_unlock_address);
+            if let Some(unlocked_alias_or_nft_address) = unlocked_alias_or_nft_address {
+                available_input_addresses.push(unlocked_alias_or_nft_address);
+            }
+        }
 
         // Try to select inputs with required inputs for utxo chains alone before requesting more inputs from addresses.
         let mut input_selection = InputSelection::new(
@@ -193,7 +198,6 @@ impl<'a> ClientBlockBuilder<'a> {
                                     *internal as u32,
                                     address_index,
                                 ])),
-                                bech32_address: str_address.clone(),
                             });
                         }
                     }