+++ client/src/api/block_builder/input_selection/mod.rs
@@ -10,6 +10,7 @@ mod manual;
 mod sender_issuer;
 mod utxo_chains;
 
+pub(crate) use self::core::is_alias_transition;
 pub use self::{
     core::{Burn, BurnDto, Error, InputSelection, Requirement, Selected},
     helpers::minimum_storage_deposit_basic_output,