+++ client/src/api/high_level.rs
@@ -25,7 +25,7 @@ use crate::{
     error::{Error, Result},
     node_api::indexer::query_parameters::QueryParameter,
     secret::SecretManager,
-    Client,
+    unix_timestamp_now, Client,
 };
 
 impl Client {
@@ -334,10 +334,7 @@ impl Client {
     /// Returns the local time checked with the timestamp of the latest milestone, if the difference is larger than 5
     /// minutes an error is returned to prevent locking outputs by accident for a wrong time.
     pub async fn get_time_checked(&self) -> Result<u32> {
-        let current_time = instant::SystemTime::now()
-            .duration_since(instant::SystemTime::UNIX_EPOCH)
-            .expect("time went backwards")
-            .as_secs() as u32;
+        let current_time = unix_timestamp_now();
 
         let network_info = self.get_network_info().await?;