+++ client/tests/input_selection/nft_outputs.rs
@@ -33,11 +33,13 @@ fn input_nft_eq_output_nft() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id_2,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -72,12 +74,14 @@ fn transition_nft_id_zero() {
         None,
         None,
         None,
+        None,
     )]);
     let nft_id = NftId::from(inputs[0].output_id());
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -107,6 +111,7 @@ fn input_amount_lt_output_amount() {
         1_000_000,
         nft_id_2,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -151,6 +156,7 @@ fn basic_output_with_nft_input() {
         2_229_500,
         nft_id_2,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -206,6 +212,7 @@ fn mint_nft() {
         None,
         None,
         None,
+        None,
     )]);
 
     let selected = InputSelection::new(
@@ -239,6 +246,7 @@ fn burn_nft() {
         2_000_000,
         nft_id_2,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -284,11 +292,13 @@ fn not_enough_storage_deposit_for_remainder() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id_2,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -337,6 +347,7 @@ fn missing_input_for_nft_output() {
         None,
         None,
         None,
+        None,
     )]);
 
     let selected = InputSelection::new(
@@ -377,6 +388,7 @@ fn missing_input_for_nft_output_but_created() {
         None,
         None,
         None,
+        None,
     )]);
 
     let selected = InputSelection::new(
@@ -405,6 +417,7 @@ fn nft_in_output_and_sender() {
             None,
             None,
             None,
+            None,
         ),
         Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
@@ -413,6 +426,7 @@ fn nft_in_output_and_sender() {
             1_000_000,
             nft_id_1,
             BECH32_ADDRESS_ED25519_0,
+            None,
             None,
             None,
             None,
@@ -466,6 +480,7 @@ fn missing_ed25519_sender() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -473,6 +488,7 @@ fn missing_ed25519_sender() {
         BECH32_ADDRESS_ED25519_0,
         None,
         Some(BECH32_ADDRESS_ED25519_1),
+        None,
         None,
         None,
         None,
@@ -516,6 +532,7 @@ fn missing_ed25519_issuer_created() {
         Some(BECH32_ADDRESS_ED25519_1),
         None,
         None,
+        None,
     )]);
 
     let selected = InputSelection::new(
@@ -546,6 +563,7 @@ fn missing_ed25519_issuer_transition() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -554,6 +572,7 @@ fn missing_ed25519_issuer_transition() {
         None,
         None,
         Some(BECH32_ADDRESS_ED25519_1),
+        None,
         None,
         None,
     )]);
@@ -583,6 +602,7 @@ fn missing_alias_sender() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -590,6 +610,7 @@ fn missing_alias_sender() {
         BECH32_ADDRESS_ED25519_0,
         None,
         Some(BECH32_ADDRESS_ALIAS_1),
+        None,
         None,
         None,
         None,
@@ -633,6 +654,7 @@ fn missing_alias_issuer_created() {
         Some(BECH32_ADDRESS_ALIAS_1),
         None,
         None,
+        None,
     )]);
 
     let selected = InputSelection::new(
@@ -663,6 +685,7 @@ fn missing_alias_issuer_transition() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -671,6 +694,7 @@ fn missing_alias_issuer_transition() {
         None,
         None,
         Some(BECH32_ADDRESS_ALIAS_1),
+        None,
         None,
         None,
     )]);
@@ -700,6 +724,7 @@ fn missing_nft_sender() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -707,6 +732,7 @@ fn missing_nft_sender() {
         BECH32_ADDRESS_ED25519_0,
         None,
         Some(BECH32_ADDRESS_NFT_1),
+        None,
         None,
         None,
         None,
@@ -750,6 +776,7 @@ fn missing_nft_issuer_created() {
         Some(BECH32_ADDRESS_NFT_1),
         None,
         None,
+        None,
     )]);
 
     let selected = InputSelection::new(
@@ -780,6 +807,7 @@ fn missing_nft_issuer_transition() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -788,6 +816,7 @@ fn missing_nft_issuer_transition() {
         None,
         None,
         Some(BECH32_ADDRESS_NFT_1),
+        None,
         None,
         None,
     )]);
@@ -818,6 +847,7 @@ fn increase_nft_amount() {
             None,
             None,
             None,
+            None,
         ),
         Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
@@ -825,6 +855,7 @@ fn increase_nft_amount() {
         3_000_000,
         nft_id_1,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -860,6 +891,7 @@ fn decrease_nft_amount() {
             None,
             None,
             None,
+            None,
         ),
         Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
@@ -867,6 +899,7 @@ fn decrease_nft_amount() {
         1_000_000,
         nft_id_1,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -914,6 +947,7 @@ fn prefer_basic_to_nft() {
             None,
             None,
             None,
+            None,
         ),
         Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
@@ -952,6 +986,7 @@ fn take_amount_from_nft_to_fund_basic() {
             2_000_000,
             nft_id_1,
             BECH32_ADDRESS_ED25519_0,
+            None,
             None,
             None,
             None,
@@ -1014,6 +1049,7 @@ fn nft_burn_should_validate_nft_sender() {
             None,
             None,
             None,
+            None,
         ),
     ]);
     let outputs = build_outputs(vec![Basic(
@@ -1057,6 +1093,7 @@ fn nft_burn_should_validate_nft_address() {
             None,
             None,
             None,
+            None,
         ),
     ]);
     let outputs = build_outputs(vec![Basic(
@@ -1093,6 +1130,7 @@ fn transitioned_zero_nft_id_no_longer_is_zero() {
         2_000_000,
         nft_id_0,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,