+++ client/tests/signing/basic.rs
@@ -89,7 +89,7 @@ async fn single_ed25519_unlock() -> Result<()> {
     };
 
     let unlocks = secret_manager
-        .sign_transaction_essence(&prepared_transaction_data)
+        .sign_transaction_essence(&prepared_transaction_data, Some(0))
         .await?;
 
     assert_eq!(unlocks.len(), 1);
@@ -211,7 +211,7 @@ async fn ed25519_reference_unlocks() -> Result<()> {
     };
 
     let unlocks = secret_manager
-        .sign_transaction_essence(&prepared_transaction_data)
+        .sign_transaction_essence(&prepared_transaction_data, Some(0))
         .await?;
 
     assert_eq!(unlocks.len(), 3);
@@ -333,7 +333,7 @@ async fn two_signature_unlocks() -> Result<()> {
     };
 
     let unlocks = secret_manager
-        .sign_transaction_essence(&prepared_transaction_data)
+        .sign_transaction_essence(&prepared_transaction_data, Some(0))
         .await?;
 
     assert_eq!(unlocks.len(), 2);